requirements.txt
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
ConfigArgParse==1.2.3
coverage==5.1
diskcache==4.1.0
entrypoints==0.3
flake8==3.7.9
idna==2.9
importlib-metadata==1.6.0
joblib==0.14.1
mccabe==0.6.1
more-itertools==8.2.0
nltk==3.5
packaging==20.3
pdfminer.six==20200402
pluggy==0.13.1
praw==7.0.0
prawcore==1.3.0
py==1.8.1
pycodestyle==2.5.0
pycryptodome==3.9.7
pyflakes==2.1.1
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.8.1
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
six==1.14.0
sortedcontainers==2.1.0
textblob==0.15.3
tqdm==4.46.0
update-checker==0.17
urllib3==1.25.9
wcwidth==0.1.9
websocket-client==0.57.0
zipp==3.1.0
